record: unexport field V.
The first idea to export this field is to make fast assignment to record.
Turn out it was bad idea which cause misused (one time I made a mistake by
assign it to value other than float64, int64, and string).

To prevent the same mistake that I made, the field now will be hidden.

Additional changes,

* NewRecord now return nil record, the old NewRecord renamed to
  NewRecordBy.
* Clone(): new method for cloning current record.
* IsNil(): new method to check if record is nil or not.
* Interface(): new method, return the interface value.
* IsEqual(): now with record as parameter, the old IsEqual has been
  renamed to IsEqualToInterface().
* Reset(): new method, which will reset the record to empty string or zero.